Created dataController and used to read data from google spreadsheet
A Google service account has been created for this app to use and read
from the Google Sheet to allow for real time updating. Currently it is stored
locally once the data has been retrieved from the sheet, but it should later
query each time a user request comes in.